refactor: LottoDrawingMachine이 하고 있던 책임을 WinningLotto로 위치 변경
